\name{ Rtemplate }
\alias{ Rtemplate }
\title{ Rtemplate Title }
\description{ Rtemplate Description }
\usage{ x <- Rtemplate(n1, n2) }
\arguments{
   \item{n1}{First Arg}
   \item{n2}{Second Arg}
}
\details{ Other details about the Rtemplate }
\value{ Return value of Rtemplate }
\references{ Rtemplate References }
\author{ Herodotos Herodotou }
\note{ Rtemplate Further notes }
\keyword{ methods }

